fix rrio optionals

+++ it/kafka/src/test/java/org/apache/beam/it/kafka/KafkaIOLT.java
@@ -145,7 +145,7 @@ public final class KafkaIOLT extends IOLoadTestBase {
               region,
               readInfo.jobId(),
               getBeamMetricsName(PipelineMetricsType.COUNTER, READ_ELEMENT_METRIC_NAME));
-      assertEquals(configuration.getNumRows(), numRecords, 10.0);
+      assertEquals((double) configuration.getNumRows(), numRecords, 10.0);
     } finally {
       // clean up pipelines
       if (pipelineLauncher.getJobStatus(project, region, writeInfo.jobId())
